Add missing columns to a data frame

add_missing_column() adds one or more columns to an existing data frame only if they do not already exist.

This is a simple wrapper around add_column().

add_missing_column(
  .data,
  ...,
  .before = NULL,
  .after = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal")
)

Arguments

.data
Data frame to append to.

...
<dynamic-dots> Name-value pairs, passed on to tibble(). All values must have the same size of .data or size 1.

.before
One-based column index or column name where to add the new columns, default: after last column.

.after
One-based column index or column name where to add the new columns, default: after last column.

.name_repair
Treatment of problematic column names:

"minimal": No name repair or checks, beyond basic existence,

"unique": Make sure names are unique and not empty,

"check_unique": (default value), no name repair, but check they are unique,

"universal": Make the names unique and syntactic

a function: apply custom name repair (e.g., .name_repair = make.names for names in the style of base R).

A purrr-style anonymous function, see rlang::as_function()

This argument is passed on as repair to vctrs::vec_as_names(). See there for more details on these terms and the strategies used to enforce them.

See also

Examples

# add new columns x and y
add_missing_column(mtcars, x = 1, y = NA)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb x  y
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4 1 NA
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4 1 NA
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1 1 NA
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 1 NA
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2 1 NA
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1 1 NA
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4 1 NA
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2 1 NA
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2 1 NA
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 1 NA
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 1 NA
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3 1 NA
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3 1 NA
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3 1 NA
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4 1 NA
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4 1 NA
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4 1 NA
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1 1 NA
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 1 NA
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1 1 NA
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1 1 NA
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2 1 NA
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2 1 NA
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4 1 NA
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2 1 NA
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1 1 NA
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2 1 NA
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 1 NA
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4 1 NA
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6 1 NA
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8 1 NA
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2 1 NA

# add new columns from a named vector
new_cols <- c(x = 1, y = NA)
add_missing_column(mtcars, !!!new_cols)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb x  y
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4 1 NA
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4 1 NA
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1 1 NA
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 1 NA
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2 1 NA
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1 1 NA
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4 1 NA
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2 1 NA
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2 1 NA
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 1 NA
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 1 NA
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3 1 NA
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3 1 NA
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3 1 NA
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4 1 NA
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4 1 NA
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4 1 NA
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1 1 NA
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 1 NA
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1 1 NA
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1 1 NA
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2 1 NA
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2 1 NA
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4 1 NA
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2 1 NA
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1 1 NA
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2 1 NA
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 1 NA
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4 1 NA
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6 1 NA
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8 1 NA
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2 1 NA