Check file specification

check_spec() will check a list of data against an expected file specification.

check_spec(
  table_list,
  table_spec,
  type_label = "Import File",
  .silent = TRUE,
  .include_table_type = NULL
)

Arguments

table_list

A list of data frames to check against the spec.

table_spec

The table spec. See details.

type_label

A character label for the check message.

.silent

Logical whether or not to print results to the console.

.include_table_type

Character vector of 'type' values from the spec 'tables' to include in the check. Useful to remove certain table types, such as enumerations. If NULL, all tables are included.

Value

A list containing the results of the check.

See also